Added react-native-vector-icons and accounting libraries to complete following tasks.
Create the last row on number pad with delete and refund buttons. The amount charged displays at the top of the screen and changes in a correct dollar amount depending on what buttons on the number pad are pressed.

When the delete is hit, amount is corrected thanks to accounting.formatMoney and dividing whatever value that is by 100.

When the refund button is pressed the state of the component is changed, and the font color for the amount appears red. Will added text when I create the header.

Next Objective is to create the header for that page and begin making more screens.