Customers can now be created and added dynamically to payment. Additionally, no more duplicate lists are being made. This was done by setting the arrays that held customers equal to null. However, is does mean the API call will need to keep being made if the user goes back and forth between screens.
Not the most efficient way to do things, but it works for now.

Next goal I'm working on is getting the tax and service fee amounts to adjust the total amount displayed on the screen dynamically as well. Keep getting a NaN as the number. Will need to spend time looking at that for my remote day.