Got custom tips used with AsyncStorage. Default tips work with a constant array that can exported and imported.
Was unable to get this same method to work with the user's custom tips. However, those do work and can be easily changeable. The tips page should be done now.

Next is to proceed with the completion of a payment. Will look at the links I received from MX Merchant to further understand how the tips should work during a payment.