The segmented control tab for the tip bar on the CollectTip.js now works if the user wants to use their custom tip percentages. Approval screen has correct payment data showing up with date and time, amount, and the authCode.
All that data is sent to SignatureScreen, to which I'm getting the correct data.

When the user is choosing what tip they want to selected the total changes dynamically.

A few bug fixes were made along the way for instance with date and time values not showing up as expecting. Making the ApprovalOverlay wider.

Finally, made some helper functions, most importantly storageGet and storageSet are now used for getting and setting AsyncStorage keys. This helps reduce the lines of code in each file significantly. Doing this with encodedUser is still needed to be done. MX Merchant's sandbox API was giving me some problems over the weekend.