Moved the screen components to their own folder with in components. Did the same for the helper methods.
On the SignatureScreen when user hits cancel payment is voided and the user is taken back to the Main screen.

There is a bug I am trying to fix now in the signature regarding the tips. Total shows up for as 0 until control is hit and also, has bugs with values not being the same. Will be looking at this later!