Layout for daily purchase history is done.
Loading times are still really bad. Would like to do something with it in the future, short lists handle better.

A loading effect could not be done at this moment because it's not issue with the speed of getting data back from the API but rather the front end having to organize that giant list it's getting back. However, for the most part that side is done.