Minor bug fixes when fresh app is installed on device.
Got rid of some console logs within the code. And minor code clean up in those files.

ViewReceiptScreen now has an overlay pop up when the user taps a "Void Refund" button confirming if the user is sure or not.

Refactored MainScreen's componentDidMount method. It's commented out within the file, will switch out the two during final build of app.

Refactored sendReceipt methods in the ReceiptScreen, and ViewReceiptScreen. Method should work as is, will need to run test on production environment. MxMerchant does not have an endpoint setup for receipts on the sandbox environment.